+++ libs/common/src/core/index.ts
@@ -1,1 +1,5 @@
+export * from './abstract-service';
+export * from './common.slack';
 export * from './config';
+export * from './document.model';
+export * from './utils';
