+++ libs/common/src/index.ts
@@ -1,6 +1,4 @@
 export * from './core';
 export * from './match';
-export * from './models';
 export * from './player';
 export * from './statistics';
-export * from './utils';
