--- libs/common/src/models/common.slack.ts
@@ -1,13 +1,0 @@
-export enum Callback {
-  FOOSBALL_MATCH = 'foosball_match',
-  UPDATE_ME = 'update_me',
-}
-
-export interface IOptions {
-  options: IOption[];
-}
-
-export interface IOption {
-  label: string;
-  value: string;
-}
