--- libs/common/src/models/document.ts
@@ -1,10 +1,0 @@
-import { serializable } from 'serializr';
-
-export interface IDocument {
-  _docId: string;
-}
-
-export abstract class FirestoreDocument implements IDocument {
-  @serializable
-  _docId!: string;
-}
