--- libs/common/src/models/index.ts
@@ -1,3 +1,0 @@
-export * from './common.slack';
-export * from './document';
-export * from './player';
