+++ libs/common/src/player/index.ts
@@ -1,2 +1,3 @@
 export * from './player.interface';
 export * from './player.service';
+export * from './player.model';
