+++ libs/common/src/statistics/stats.service.ts
@@ -2,8 +2,8 @@ import { FieldValue } from 'firebase-admin/firestore';
 
 import { CoreService } from '../core/abstract-service';
 import { IMatchResult } from '../match/match-result.model';
-import { checkFlawlessVictory, checkSuckerPunch } from '../utils';
-import { Collection } from '../utils/firestore-db';
+import { checkFlawlessVictory, checkSuckerPunch } from '../core/utils';
+import { Collection } from '../core/utils/firestore-db';
 import { IMetrics } from './metrics';
 
 const PLAYERS_COLLECTION = Collection.PLAYERS;
