+++ libs/common/src/statistics/stats.util.ts
@@ -1,6 +1,6 @@
 import { IPlayerStats } from '../player/player.interface';
 import { IMatchResult } from '../match/match-result.model';
-import { checkFlawlessVictory, checkSuckerPunch } from '../utils/various';
+import { checkFlawlessVictory, checkSuckerPunch } from '../core/utils/various.utils';
 
 enum EntityType {
   PLAYER,
