--- libs/common/src/utils/index.ts
@@ -1,3 +1,0 @@
-export * from './firestore-db';
-export * from './gemini.helper';
-export * from './various';
