--- libs/common/src/utils/serializers/string-array.serializer.ts
@@ -1,7 +1,0 @@
-export function stringArraySerializer(v: any) {
-  return v;
-}
-
-export function stringArrayDeserializer(jsonValue: any, context: any, oldValue: any) {
-  return jsonValue;
-}
